Rework changes for queueSize
